test: use a WorkbenchTestCase
